moved CryptoError from enclave-ffi-types to wasmi_runtime::crypto::errors
