+++ cosmwasm/packages/enclave-ffi-types/src/lib.rs
@@ -4,8 +4,7 @@
 mod types;
 
 pub use types::{
-    CryptoError, Ctx, EnclaveBuffer, EnclaveError, HandleResult, InitResult, QueryResult,
-    UserSpaceBuffer,
+    Ctx, EnclaveBuffer, EnclaveError, HandleResult, InitResult, QueryResult, UserSpaceBuffer,
 };
 
 pub const ENCRYPTED_SEED_SIZE: usize = 48;
