+++ cosmwasm/packages/enclave-ffi-types/src/types.rs
@@ -109,83 +109,6 @@ pub enum EnclaveError {
     Panic,
 }
 
-#[repr(C)]
-#[derive(Debug, Display)]
-pub enum CryptoError {
-    /// The `DerivingKeyError` error.
-    ///
-    /// This error means that the ECDH process failed.
-    DerivingKeyError,
-    // {
-    //     self_key: [u8; 64],
-    //     other_key: [u8; 64],
-    // },
-    /// The `MissingKeyError` error.
-    ///
-    /// This error means that a key was missing.
-    MissingKeyError,
-    //  {
-    //     key_type: &'static str,
-    // },
-    /// The `DecryptionError` error.
-    ///
-    /// This error means that the symmetric decryption has failed for some reason.
-    DecryptionError,
-    /// The `ImproperEncryption` error.
-    ///
-    /// This error means that the ciphertext provided was imporper.
-    /// e.g. MAC wasn't valid, missing IV etc.
-    ImproperEncryption,
-    /// The `EncryptionError` error.
-    ///
-    /// This error means that the symmetric encryption has failed for some reason.
-    EncryptionError,
-    /// The `SigningError` error.
-    ///
-    /// This error means that the signing process has failed for some reason.
-    SigningError,
-    // {
-    //     hashed_msg: [u8; 32],
-    // },
-    /// The `ParsingError` error.
-    ///
-    /// This error means that the signature couldn't be parsed correctly.
-    ParsingError,
-    //  {
-    //     sig: [u8; 65],
-    // },
-    /// The `RecoveryError` error.
-    ///
-    /// This error means that the public key can't be recovered from that message & signature.
-    RecoveryError,
-    //  {
-    //     sig: [u8; 65],
-    // },
-    /// The `KeyError` error.
-    ///
-    /// This error means that a key wasn't vaild.
-    /// e.g. PrivateKey, PubliKey, SharedSecret.
-    // #[cfg(feature = "asymmetric")]
-    KeyError,
-    //  {
-    //     key_type: &'static str,
-    //     err: Option<secp256k1::Error>,
-    // },
-    // #[cfg(not(feature = "asymmetric"))]
-    // KeyError { key_type: &'static str, err: Option<()> },
-    // /// The `RandomError` error.
-    // ///
-    // /// This error means that the random function had failed generating randomness.
-    // #[cfg(feature = "std")]
-    // RandomError {
-    //     err: rand::Error,
-    // },
-    // #[cfg(feature = "sgx")]
-    RandomError, // {
-                 //     err: sgx_types::sgx_status_t,
-                 // }
-}
-
 /// This struct is returned from ecall_init.
 /// cbindgen:prefix-with-name
 #[repr(C)]
