+++ cosmwasm/packages/wasmi-runtime/src/crypto/ed25519.rs
@@ -5,7 +5,7 @@ use sgx_types::sgx_align_ec256_private_t;
 
 // use x25519_dalek;
 
-use enclave_ffi_types::CryptoError;
+use crate::crypto::errors::CryptoError;
 
 use super::rng::rand_slice;
