+++ cosmwasm/packages/wasmi-runtime/src/crypto/rng.rs
@@ -1,4 +1,4 @@
-use enclave_ffi_types::CryptoError;
+use crate::crypto::errors::CryptoError;
 use sgx_trts::trts::rsgx_read_rand;
 
 pub fn rand_slice(rand: &mut [u8]) -> Result<(), CryptoError> {
