+++ cosmwasm/packages/wasmi-runtime/src/crypto/sha.rs
@@ -17,7 +17,7 @@ pub fn sha_256(data: &[u8]) -> [u8; HASH_SIZE] {
 //         Keychain, CONSENSUS_SEED_SEALING_PATH, KEY_MANAGER, REGISTRATION_KEY_SEALING_PATH,
 //     };
 //     use crate::crypto::{Kdf, KeyPair, Seed};
-//     use enclave_ffi_types::CryptoError;
+//     use crate::crypto::errors::CryptoError;
 //
 //     // todo: fix test vectors to actually work
 //     fn test_derive_key() {
