+++ cosmwasm/packages/wasmi-runtime/src/crypto/traits.rs
@@ -1,7 +1,10 @@
-use enclave_ffi_types::{CryptoError, EnclaveError};
+use enclave_ffi_types::EnclaveError;
+
+use crate::crypto::errors::CryptoError;
 
 pub const HMAC_SIGNATURE_SIZE: usize = 32;
 pub const EC_256_PRIVATE_KEY_SIZE: usize = 32;
+
 pub trait Encryptable {
     fn encrypt(&self, plaintext: &[u8]) -> Result<Vec<u8>, CryptoError>;
     fn decrypt(&self, ciphertext: &[u8]) -> Result<Vec<u8>, CryptoError>;
